Exception handler refactoring, dal fix
