+++ app/main.py
@@ -1,12 +1,11 @@
 import uvicorn
-from fastapi import FastAPI, Request
-from fastapi.responses import JSONResponse
+from fastapi import FastAPI
 
 from app.api.book.router import router as book_router
 from app.api.customer.router import router as customer_router
 from app.api.employee.router import router as employee_router
-from app.api.exceptions import LibraryApiException
 from app.api.order.router import router as order_router
+from .exception_handler import init_exception_handlers
 from .logger_config import logger
 
 app = FastAPI()
@@ -15,22 +14,7 @@ app.include_router(book_router)
 app.include_router(customer_router)
 app.include_router(employee_router)
 app.include_router(order_router)
-
-
-@app.exception_handler(LibraryApiException)
-async def library_api_exception_handler(_: Request, exc: LibraryApiException):
-    detail = {"message": "api error"}
-    if exc.message:
-        detail["message"] = exc.message
-
-    if exc.error_name:
-        detail["message"] = f"<{exc.error_name}> - {detail['message']}"
-
-    logger.error(exc)
-    return JSONResponse(
-        status_code=exc.status_code, content={"detail": detail["message"]}
-    )
-
+init_exception_handlers(app)
 
 if __name__ == "__main__":
     logger.debug("Starting app")
